cleaned up code and remove unused modules
