+++ .gitignore
@@ -1,6 +1,5 @@
 .DS_Store
 node_modules/
-archive/
 dist/
 spec.txt
 .vscode/
