+++ lib/global.js
@@ -5,14 +5,14 @@ var mkdirp = require('mkdirp');
 
 // Define Base folders 
 var srcDir = {
-    pages: '../views/pages/',
-    posts: '../views/posts/',
-    drafts: '../views/drafts/',
-    layouts: '../views/layouts/',
-    includes: '../public/src/js/',
-    sass: '../public/src/sass/*.scss',
-    dist: '../dist/',
-    distCss: '../dist/css/'
+    pages: '../../../views/pages/',
+    posts: '../../../views/posts/',
+    drafts: '../../../views/drafts/',
+    layouts: '../../../views/layouts/',
+    includes: '../../../public/src/js/',
+    sass: '../../../public/src/sass/*.scss',
+    dist: '../../../dist/',
+    distCss: '../../../dist/css/'
 };
 
 // siteData will hold the meta information for the site
