+++ lib/tasks/index.js
@@ -1,0 +1,9 @@
+var post = require('./post/post');
+var page = require('./page/page');
+var build = require('./build/build');
+
+module.exports = {
+    post: post,
+    page: page,
+    build: build
+};
