+++ lib/tasks/page/page.js
@@ -6,22 +6,13 @@ var gulp = require('gulp');
 var mkdirp = require('mkdirp');
 var moment = require('moment');
 
-var global = require('../../lib/global');
+var global = require('../../global');
 
-module.exports = function (gulp) {
+function page() {
 
-    // Gulp Task for generating a new page
-    gulp.task('page', function () {
-
-        // If the output is either -t or --title  then we store this to title
-        if (argv.title || argv.t) {
-            var title = argv.title || argv.t;
-        } else {
-
-            // Otherwise we log the error out to console if user didn't input -t or --title
-            return 'Error: You must enter a title for your page. Use --title "Page title" or -t "Page title"';
-        }
-
+    // If the output is either -t or --title  then we store this to title
+    if (argv.title || argv.t) {
+        var title = argv.title || argv.t;
         var slug = (argv.slug || argv.s) ? global.hyphenateSlug(argv.slug || argv.s) : global.hyphenateSlug(title.toString());
 
         // Filename has the value of the file path of the page in question
@@ -68,5 +59,16 @@ module.exports = function (gulp) {
                 });
             });
         });
-    });
+
+    } else {
+
+        // Otherwise we log the error out to console if user didn't input -t or --title
+        return 'Error: You must enter a title for your page. Use --title "Page title" or -t "Page title"';
+    }
+
+}
+
+
+module.exports = {
+    page: page
 };
