--- tasks/index.js
@@ -1,11 +1,0 @@
-module.exports = function(gulp) {
-    // global.js
-    require('../lib/global'),
-    // post.js
-    require('./post/post')(gulp);
-    // page.js
-    require('./page/page')(gulp);
-    // build.js
-    require('./build/build')(gulp);
-
-};
